+++ magma_smaht/commands/create_meta_workflow_run.py
@@ -6,11 +6,21 @@ from magma_smaht.create_metawfr import (
     mwfr_fastqc,
     mwfr_hic_alignment,
     mwfr_ont_alignment,
+    mwfr_cram_to_fastq_paired_end,
+    mwfr_bamqc_short_read,
+    mwfr_ubam_qc_long_read,
 )
 from magma_smaht.utils import get_auth_key
 
 
-@click.command()
+@click.group()
+@click.help_option("--help", "-h")
+def cli():
+    # create group for all the commands. -h will show all available commands
+    pass
+
+
+@cli.command()
 @click.help_option("--help", "-h")
 @click.option(
     "-f", "--fileset-accession", required=True, type=str, help="Fileset accession"
@@ -29,13 +39,13 @@ from magma_smaht.utils import get_auth_key
     type=str,
     help="Name of environment in smaht-keys file",
 )
-def mwfr_illumina_alignment_cmd(fileset_accession, length_required, auth_env):
-    """Creates a MetaWorflowRun item in the portal for Illumina alignment of submitted files within a fileset"""
+def align_illumina(fileset_accession, length_required, auth_env):
+    """Alignment MWFR for Illumina data"""
     smaht_key = get_auth_key(auth_env)
     mwfr_illumina_alignment(fileset_accession, length_required, smaht_key)
 
 
-@click.command()
+@cli.command()
 @click.help_option("--help", "-h")
 @click.option(
     "-f", "--fileset-accession", required=True, type=str, help="Fileset accession"
@@ -47,13 +57,13 @@ def mwfr_illumina_alignment_cmd(fileset_accession, length_required, auth_env):
     type=str,
     help="Name of environment in smaht-keys file",
 )
-def mwfr_pacbio_alignment_cmd(fileset_accession, auth_env):
-    """Creates a MetaWorflowRun item in the portal for PacBio alignment of submitted files within a fileset"""
+def align_pacbio(fileset_accession, auth_env):
+    """Alignment MWFR for PacBio data"""
     smaht_key = get_auth_key(auth_env)
     mwfr_pacbio_alignment(fileset_accession, smaht_key)
 
 
-@click.command()
+@cli.command()
 @click.help_option("--help", "-h")
 @click.option(
     "-f", "--fileset-accession", required=True, type=str, help="Fileset accession"
@@ -65,13 +75,13 @@ def mwfr_pacbio_alignment_cmd(fileset_accession, auth_env):
     type=str,
     help="Name of environment in smaht-keys file",
 )
-def mwfr_hic_alignment_cmd(fileset_accession, auth_env):
-    """Creates a MetaWorflowRun item in the portal for HIC alignment of submitted files within a fileset"""
+def align_hic(fileset_accession, auth_env):
+    """Alignment MWFR for HIC data"""
     smaht_key = get_auth_key(auth_env)
     mwfr_hic_alignment(fileset_accession, smaht_key)
 
 
-@click.command()
+@cli.command()
 @click.help_option("--help", "-h")
 @click.option(
     "-f", "--fileset-accession", required=True, type=str, help="Fileset accession"
@@ -83,14 +93,13 @@ def mwfr_hic_alignment_cmd(fileset_accession, auth_env):
     type=str,
     help="Name of environment in smaht-keys file",
 )
-def mwfr_ont_alignment_cmd(fileset_accession, auth_env):
-    """Creates a MetaWorflowRun item in the portal for ONT alignment of submitted files within a fileset"""
+def align_ont(fileset_accession, auth_env):
+    """Alignment MWFR for ONT data"""
     smaht_key = get_auth_key(auth_env)
     mwfr_ont_alignment(fileset_accession, smaht_key)
 
 
-
-@click.command()
+@cli.command()
 @click.help_option("--help", "-h")
 @click.option(
     "-f", "--fileset-accession", required=True, type=str, help="Fileset accession"
@@ -102,6 +111,63 @@ def mwfr_ont_alignment_cmd(fileset_accession, auth_env):
     type=str,
     help="Name of environment in smaht-keys file",
 )
-def mwfr_fastqc_cmd(fileset_accession, auth_env):
+def qc_short_read_fastq(fileset_accession, auth_env):
+    """QC MWFR for short-read FASTQs"""
     smaht_key = get_auth_key(auth_env)
     mwfr_fastqc(fileset_accession, smaht_key)
+
+
+@cli.command()
+@click.help_option("--help", "-h")
+@click.option(
+    "-f", "--fileset-accession", required=True, type=str, help="Fileset accession"
+)
+@click.option(
+    "-e",
+    "--auth-env",
+    required=True,
+    type=str,
+    help="Name of environment in smaht-keys file",
+)
+def qc_long_read_ubam(fileset_accession, auth_env):
+    """QC MWFR for unaligned long-read BAMs"""
+    smaht_key = get_auth_key(auth_env)
+    mwfr_ubam_qc_long_read(fileset_accession, smaht_key)
+
+
+@cli.command()
+@click.help_option("--help", "-h")
+@click.option("-f", "--file-accession", required=True, type=str, help="File accession")
+@click.option(
+    "-e",
+    "--auth-env",
+    required=True,
+    type=str,
+    help="Name of environment in smaht-keys file",
+)
+def qc_short_read_bam(file_accession, auth_env):
+    """QC MWFR for aligned short-read BAMs"""
+    smaht_key = get_auth_key(auth_env)
+    mwfr_bamqc_short_read(file_accession, smaht_key)
+
+
+@cli.command()
+@click.help_option("--help", "-h")
+@click.option(
+    "-f", "--fileset-accession", required=True, type=str, help="Fileset accession"
+)
+@click.option(
+    "-e",
+    "--auth-env",
+    required=True,
+    type=str,
+    help="Name of environment in smaht-keys file",
+)
+def conversion_cram_to_fastq(fileset_accession, auth_env):
+    """Conversion MWFR for CRAM to FASTQ (paired-end)"""
+    smaht_key = get_auth_key(auth_env)
+    mwfr_cram_to_fastq_paired_end(fileset_accession, smaht_key)
+
+
+if __name__ == "__main__":
+    cli()